Support conditional deps using "depends on A if B"
Extend the "depends on" syntax to support conditional dependencies
using "depends on A if B". While functionally equivalent to "depends
on !B || A", "depends on A if B" is much more readable.

This change is implemented by converting the "A if B" syntax into the
"!B || A" syntax during "depends on" token processing.

Signed-off-by: Graham Roff <grahamr@qti.qualcomm.com>